Changed module organisation for event_handler
